# backend/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.2.0",
    "title": "SCA Backend",
    "description": "Sistema de controle de acesso"
  },
  "host": "localhost:3100",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "name": "Authentication",
      "in": "header"
    },
    "accessLevel": {
      "type": "apiKey",
      "name": "access-level",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/posto": {
      "get": {
        "tags": [
          "Posto"
        ],
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "nome",
            "in": "query",
            "type": "string"
          },
          {
            "name": "nivel_acesso",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Posto"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "example": "any"
                },
                "nivel_acesso": {
                  "example": "any"
                },
                "ativo_posto": {
                  "example": "any"
                },
                "sinc_posto": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "accessLevel": []
    }
  ]
}